Divide context menu items into two components
