+++ packages/ketcher-react/src/script/ui/views/components/ContextMenu/contextMenu.types.ts
@@ -1,0 +1,6 @@
+export type ContextMenuItemProps = {
+  selected: boolean
+  ci: any
+}
+
+export type ContextMenuItemData = unknown
